requirements.txt
flask
flask-cors
python-dotenv
pillow
groq
google-genai
requests
gunicorn